+++ src/integrationTest/java/uk/gov/hmcts/reform/blobrouter/data/ReconciliationReportRepositoryTest.java
@@ -8,22 +8,33 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.ReconciliationReportRepository;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.model.NewReconciliationReport;
+import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.model.ReconciliationContent;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.model.ReconciliationReport;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.statements.SupplierStatementRepository;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.statements.model.NewEnvelopeSupplierStatement;
 
+import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.UUID;
 import javax.validation.ClockProvider;
 
+import static java.time.LocalDate.now;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 
 @ActiveProfiles({"integration-test", "db-test"})
 @SpringBootTest
 public class ReconciliationReportRepositoryTest {
+
+    private static final String ACCOUNT = "account";
+    private static final String VERSION = "v1";
+    private static final NewEnvelopeSupplierStatement NEW_STATEMENT = new NewEnvelopeSupplierStatement(
+        now(),
+        "{\"content\":\"some_content\"}",
+        "supplier version"
+    );
 
     @Autowired SupplierStatementRepository statementRepo;
     @Autowired ReconciliationReportRepository reportRepo;
@@ -116,5 +127,77 @@ public class ReconciliationReportRepositoryTest {
 
         // then
         assertThat(report).isEmpty();
+    }
+
+    @Test
+    void should_not_find_anything_when_db_is_empty() {
+        // when
+        Optional<ReconciliationContent> report = reportRepo.getReconciliationReport(now(), ACCOUNT);
+
+        // then
+        assertThat(report).isEmpty();
+    }
+
+    @Test
+    void should_not_find_anything_when_conditions_do_not_match() {
+        // given
+        saveNewReportsAndGetLastId("{}");
+
+        // when
+        Optional<ReconciliationContent> report = reportRepo.getReconciliationReport(now().minusDays(1), ACCOUNT);
+
+        // then
+        assertThat(report).isEmpty();
+    }
+
+    @Test
+    void should_find_a_report_when_conditions_are_met() {
+        // given
+        var expectedReportContent = "{}";
+        var id = saveNewReportsAndGetLastId(expectedReportContent);
+
+        // when
+        Optional<ReconciliationContent> report = reportRepo.getReconciliationReport(now(), ACCOUNT);
+
+        // then
+        assertThat(report)
+            .isNotEmpty()
+            .get()
+            .usingRecursiveComparison()
+            .isEqualTo(new ReconciliationContent(id, expectedReportContent, VERSION));
+    }
+
+    @Test
+    void should_find_only_latest_report_when_conditions_are_met() {
+        // given
+        var expectedReportContent = "[]";
+        var id = saveNewReportsAndGetLastId("{}", expectedReportContent);
+
+        // when
+        Optional<ReconciliationContent> report = reportRepo.getReconciliationReport(now(), ACCOUNT);
+
+        // then
+        assertThat(report)
+            .isNotEmpty()
+            .get()
+            .usingRecursiveComparison()
+            .isEqualTo(new ReconciliationContent(id, expectedReportContent, VERSION));
+    }
+
+    // Only last ID/report matters
+    private UUID saveNewReportsAndGetLastId(String... reportContents) {
+        UUID lastId = null;
+
+        try {
+            for (String contents : reportContents) {
+                var statementId = statementRepo.save(NEW_STATEMENT);
+                var report = new NewReconciliationReport(statementId, ACCOUNT, contents, VERSION);
+                lastId = reportRepo.save(report);
+            }
+        } catch (SQLException exception) {
+            throw new RuntimeException("Could not save reports for testing", exception);
+        }
+
+        return lastId;
     }
 }
