+++ src/main/java/uk/gov/hmcts/reform/blobrouter/data/reconciliation/reports/ReconciliationReportRepository.java
@@ -5,10 +5,12 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.model.NewReconciliationReport;
+import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.model.ReconciliationContent;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.model.ReconciliationReport;
 
 import java.sql.SQLException;
 import java.time.Clock;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.UUID;
@@ -21,15 +23,18 @@ public class ReconciliationReportRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final ReconciliationReportRowMapper rowMapper;
+    private final ReconciliationContentMapper contentMapper;
     private final Clock clock;
 
     public ReconciliationReportRepository(
         NamedParameterJdbcTemplate jdbcTemplate,
         ReconciliationReportRowMapper rowMapper,
+        ReconciliationContentMapper contentMapper,
         ClockProvider clockProvider
     ) {
         this.jdbcTemplate = jdbcTemplate;
         this.rowMapper = rowMapper;
+        this.contentMapper = contentMapper;
         this.clock = clockProvider.getClock();
     }
 
@@ -59,6 +64,27 @@ public class ReconciliationReportRepository {
                 this.rowMapper
             );
             return Optional.of(report);
+        } catch (EmptyResultDataAccessException ex) {
+            return Optional.empty();
+        }
+    }
+
+    public Optional<ReconciliationContent> getReconciliationReport(LocalDate forDate, String account) {
+        try {
+            ReconciliationContent report = jdbcTemplate.queryForObject(
+                "SELECT id, content, content_type_version "
+                    + "FROM envelope_reconciliation_reports "
+                    + "WHERE account = :account"
+                    + "  AND DATE(created_at) = :date "
+                    + "ORDER BY created_at DESC "
+                    + "LIMIT 1",
+                new MapSqlParameterSource()
+                    .addValue("date", forDate)
+                    .addValue("account", account),
+                contentMapper
+            );
+
+            return Optional.ofNullable(report);
         } catch (EmptyResultDataAccessException ex) {
             return Optional.empty();
         }
