+++ src/main/java/uk/gov/hmcts/reform/blobrouter/data/reports/ReportRepository.java
@@ -1,15 +1,13 @@
 package uk.gov.hmcts.reform.blobrouter.data.reports;
 
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
+import uk.gov.hmcts.reform.blobrouter.data.reconciliation.reports.ReconciliationContentMapper;
 
 import java.sql.Timestamp;
 import java.time.Instant;
-import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public class ReportRepository {
@@ -43,26 +41,5 @@ public class ReportRepository {
                 .addValue("to", Timestamp.from(to)),
             this.mapper
         );
-    }
-
-    public Optional<ReconciliationContent> getReconciliationReport(LocalDate forDate, String account) {
-        try {
-            ReconciliationContent report = jdbcTemplate.queryForObject(
-                "SELECT id, content, content_type_version "
-                    + "FROM envelope_reconciliation_reports "
-                    + "WHERE account = :account"
-                    + "  AND DATE(created_at) = :date "
-                    + "ORDER BY created_at DESC "
-                    + "LIMIT 1",
-                new MapSqlParameterSource()
-                    .addValue("date", forDate)
-                    .addValue("account", account),
-                reconciliationMapper
-            );
-
-            return Optional.ofNullable(report);
-        } catch (EmptyResultDataAccessException ex) {
-            return Optional.empty();
-        }
     }
 }
